Passing netty ByteBuf from NettyRequest to ByteBufferAsyncWritableChannel (#1409)
Passing netty ByteBuf in NettyRequest to ByteBufferAsyncWritableChannel.
By doing this, we can avoid creating a java ByteBuffer ( one less memory allocation) and copying data from httpContent.content() to newly created java ByteBuffer (one less memory copy).

+++ ambry-api/src/main/java/com.github.ambry/router/AsyncWritableChannel.java
@@ -70,7 +70,7 @@ public interface AsyncWritableChannel extends Channel {
       if (result != 0) {
         src.readerIndex(src.readerIndex() + (int) result.longValue());
       }
-      futureResult.done(result.longValue(), exception);
+      futureResult.done(result, exception);
       if (callback != null) {
         callback.onCompletion(result, exception);
       }
